add return ark transaction

+++ src/main/java/io/ark/ark_client/Transaction.java
@@ -2,7 +2,6 @@ package io.ark.ark_client;
 
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.util.List;
 
 @Data
@@ -15,8 +14,8 @@ public class Transaction {
     private String senderPublicKey;
     private String senderId;
     private String recipientId;
-    private BigInteger amount;
-    private BigInteger fee;
+    private Long amount;
+    private Long fee;
     private String signature;
     private List<String> signatures;
     private Integer confirmations;
